+++ docs/CHANGELOG.md
@@ -9,13 +9,13 @@ Released on 12 February 2025<br>
 ## Version 1.1.0
 Build 1001<br>
 Released on 12 February 2025<br>
-- Added 'ac version' & 'ac --version' command to see Azure Core current version.
-- Added 'ac changelog' command to see the full Azure Core changelog.
+- Added `'ac version'` & `'ac --version'` command to see Azure Core current version.
+- Added `'ac changelog'` command to see the full Azure Core changelog.
 
 ## Version 1.2.0
 Build 1002<br>
 Released on 13 February 2025<br>
-- Added 'ac checkupdates' command for users to check if there are any new updates and update their system on the spot.
+- Added `'ac checkupdates'` command for users to check if there are any new updates and update their system on the spot.
 
 ## Version 1.2.1
 Build 1003<br>
@@ -25,16 +25,32 @@ Released on 13 February 2025<br>
 ## Version 1.3.0
 Build 1004<br>
 Released on 14 February 2025<br>
-- Removed automatic updating when running 'ac checkupdates'. Users now have to pull new updates from GitHub themselves.
+- Removed automatic updating when running `'ac checkupdates'`. Users now have to pull new updates from GitHub themselves.
 - Fixed previous changelog commands to say the prefix in the command.
 
 ## Version 1.4.0
 Build 1005<br>
 Released on 14 February 2025<br>
-- Made 'ac changelog' command look nicer.
+- Made `'ac changelog'` command look nicer.
 
 ## Version 1.4.1
 Build 1006<br>
 Released on 17 February 2025<br>
 - Fixed [README.md](README.md) contributer section to link to the respective contributers social media.
 - Added what each contributer contributed.
+
+## Version 1.5.0
+Build 1007<br>
+Released on 17 February 2025<br>
+- Added `'ac sysinfo'`, `'ac ls'`, `'ac tasklist'`, `'ac ipconfig'`, `'ac help'` & `'-p help'` commands.
+- Made `'ac sysinfo'` command nicer looking.
+
+## Version 1.5.1
+Build 1008<br>
+Released on 17 February 2025<br>
+- Made [CHANGELOG.md](CHANGELOG.md) file commands look nicer.
+
+## Version 1.5.2
+Build 1009<br>
+Released on 18 February 2025<br>
+- Fixed `'ac Restart-System'` & `'ac Shutdown-System'` not working.